+++ cmd/ocsp-responder/main.go
@@ -29,7 +29,7 @@ import (
 	"github.com/letsencrypt/boulder/issuance"
 	blog "github.com/letsencrypt/boulder/log"
 	"github.com/letsencrypt/boulder/metrics/measured_http"
-	bocsp "github.com/letsencrypt/boulder/ocsp"
+	ocsp_responder "github.com/letsencrypt/boulder/ocsp/responder"
 	"github.com/letsencrypt/boulder/rocsp"
 	rocsp_config "github.com/letsencrypt/boulder/rocsp/config"
 	"github.com/letsencrypt/boulder/sa"
@@ -106,7 +106,7 @@ func newSourceMetrics(stats prometheus.Registerer) *sourceMetrics {
 // the requirements of an OCSP request, or nil if the request should be handled.
 func (f *ocspFilter) checkRequest(req *ocsp.Request) error {
 	if req.HashAlgorithm != f.issuerKeyHashAlgorithm {
-		return fmt.Errorf("Request ca key hash using unsupported algorithm %s: %w", req.HashAlgorithm, bocsp.ErrNotFound)
+		return fmt.Errorf("Request ca key hash using unsupported algorithm %s: %w", req.HashAlgorithm, ocsp_responder.ErrNotFound)
 	}
 	// Check that this request is for the proper CA. We only iterate over
 	// issuerKeyHashes here because it is guaranteed to have the same values
@@ -118,7 +118,7 @@ func (f *ocspFilter) checkRequest(req *ocsp.Request) error {
 		}
 	}
 	if !match {
-		return fmt.Errorf("Request intended for wrong issuer cert %s: %w", hex.EncodeToString(req.IssuerKeyHash), bocsp.ErrNotFound)
+		return fmt.Errorf("Request intended for wrong issuer cert %s: %w", hex.EncodeToString(req.IssuerKeyHash), ocsp_responder.ErrNotFound)
 	}
 
 	serialString := core.SerialToString(req.SerialNumber)
@@ -130,7 +130,7 @@ func (f *ocspFilter) checkRequest(req *ocsp.Request) error {
 			}
 		}
 		if !match {
-			return fmt.Errorf("Request serial has wrong prefix: %w", bocsp.ErrNotFound)
+			return fmt.Errorf("Request serial has wrong prefix: %w", ocsp_responder.ErrNotFound)
 		}
 	}
 
@@ -182,7 +182,7 @@ type dbSelector interface {
 	WithContext(ctx context.Context) gorp.SqlExecutor
 }
 
-// Response implements the `responder.Source` interface and is called by
+// Response implements the `ocsp_responder.Source` interface and is called by
 // the HTTP server to handle a new OCSP request.
 func (src *dbSource) Response(ctx context.Context, req *ocsp.Request) ([]byte, http.Header, error) {
 	err := src.filter.checkRequest(req)
@@ -239,7 +239,7 @@ func (src *dbSource) Response(ctx context.Context, req *ocsp.Request) ([]byte, h
 		return nil, nil, fmt.Errorf("looking up OCSP response for serial: %s err: %w", serialString, ctx.Err())
 	case primaryResult := <-primaryChan:
 		if primaryResult.err != nil {
-			if !errors.Is(primaryResult.err, bocsp.ErrNotFound) {
+			if !errors.Is(primaryResult.err, ocsp_responder.ErrNotFound) {
 				src.metrics.ocspLookups.WithLabelValues("mysql_failed").Inc()
 				src.metrics.sourceUsed.WithLabelValues("error_returned").Inc()
 			}
@@ -276,7 +276,7 @@ func (src *dbSource) Response(ctx context.Context, req *ocsp.Request) ([]byte, h
 
 		// Check for error returned from the mysql lookup, return on error.
 		if primaryResult.err != nil {
-			if !errors.Is(primaryResult.err, bocsp.ErrNotFound) {
+			if !errors.Is(primaryResult.err, ocsp_responder.ErrNotFound) {
 				src.metrics.ocspLookups.WithLabelValues("mysql_failed").Inc()
 				src.metrics.sourceUsed.WithLabelValues("error_returned").Inc()
 			}
@@ -364,7 +364,7 @@ func (src dbReceiver) getResponse(ctx context.Context, req *ocsp.Request) chan l
 		certStatus, err := sa.SelectCertificateStatus(src.dbMap.WithContext(ctx), serialString)
 		if err != nil {
 			if db.IsNoRows(err) {
-				responseChan <- lookupResponse{nil, bocsp.ErrNotFound}
+				responseChan <- lookupResponse{nil, ocsp_responder.ErrNotFound}
 				return
 			}
 			src.log.AuditErrf("Looking up OCSP response in DB: %s", err)
@@ -375,15 +375,15 @@ func (src dbReceiver) getResponse(ctx context.Context, req *ocsp.Request) chan l
 
 		if certStatus.IsExpired {
 			src.log.Infof("OCSP Response not sent (expired) for CA=%s, Serial=%s", hex.EncodeToString(req.IssuerKeyHash), serialString)
-			responseChan <- lookupResponse{nil, bocsp.ErrNotFound}
+			responseChan <- lookupResponse{nil, ocsp_responder.ErrNotFound}
 			return
 		} else if certStatus.OCSPLastUpdated.IsZero() {
 			src.log.Warningf("OCSP Response not sent (ocspLastUpdated is zero) for CA=%s, Serial=%s", hex.EncodeToString(req.IssuerKeyHash), serialString)
-			responseChan <- lookupResponse{nil, bocsp.ErrNotFound}
+			responseChan <- lookupResponse{nil, ocsp_responder.ErrNotFound}
 			return
 		} else if !src.filter.responseMatchesIssuer(req, certStatus) {
 			src.log.Warningf("OCSP Response not sent (issuer and serial mismatch) for CA=%s, Serial=%s", hex.EncodeToString(req.IssuerKeyHash), serialString)
-			responseChan <- lookupResponse{nil, bocsp.ErrNotFound}
+			responseChan <- lookupResponse{nil, ocsp_responder.ErrNotFound}
 			return
 		}
 		responseChan <- lookupResponse{certStatus.OCSPResponse, err}
@@ -484,7 +484,7 @@ as generated by Boulder's ceremony command.
 	logger.Info(cmd.VersionString())
 
 	config := c.OCSPResponder
-	var source bocsp.Source
+	var source ocsp_responder.Source
 
 	if strings.HasPrefix(config.Source, "file:") {
 		url, err := url.Parse(config.Source)
@@ -495,7 +495,7 @@ as generated by Boulder's ceremony command.
 		if filename == "" {
 			filename = url.Opaque
 		}
-		source, err = bocsp.NewMemorySourceFromFile(filename, logger)
+		source, err = ocsp_responder.NewMemorySourceFromFile(filename, logger)
 		cmd.FailOnError(err, fmt.Sprintf("Couldn't read file: %s", url.Path))
 	} else {
 		// For databases, DBConfig takes precedence over Source, if present.
@@ -599,8 +599,8 @@ func (om *ocspMux) Handler(_ *http.Request) (http.Handler, string) {
 	return om.handler, "/"
 }
 
-func mux(stats prometheus.Registerer, responderPath string, source bocsp.Source, logger blog.Logger) http.Handler {
-	stripPrefix := http.StripPrefix(responderPath, bocsp.NewResponder(source, stats, logger))
+func mux(stats prometheus.Registerer, responderPath string, source ocsp_responder.Source, logger blog.Logger) http.Handler {
+	stripPrefix := http.StripPrefix(responderPath, ocsp_responder.NewResponder(source, stats, logger))
 	h := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		if r.Method == "GET" && r.URL.Path == "/" {
 			w.Header().Set("Cache-Control", "max-age=43200") // Cache for 12 hours
