+++ cmd/ocsp-updater/main.go
@@ -14,8 +14,8 @@ import (
 	"github.com/letsencrypt/boulder/cmd"
 	"github.com/letsencrypt/boulder/features"
 	bgrpc "github.com/letsencrypt/boulder/grpc"
-	"github.com/letsencrypt/boulder/ocsp_updater"
-	ocsp_updater_config "github.com/letsencrypt/boulder/ocsp_updater/config"
+	ocsp_updater "github.com/letsencrypt/boulder/ocsp/updater"
+	ocsp_updater_config "github.com/letsencrypt/boulder/ocsp/updater/config"
 	"github.com/letsencrypt/boulder/rocsp"
 	rocsp_config "github.com/letsencrypt/boulder/rocsp/config"
 	"github.com/letsencrypt/boulder/sa"
